RD-2504 added Application and UI fix that causes to sample app crash
