+++ app/build.gradle
@@ -23,12 +23,12 @@ dependencies {
     testCompile 'junit:junit:4.12'
     compile 'com.android.support:appcompat-v7:25.1.1'
     compile 'com.appsflyer:af-android-sdk:4.7.0@aar'
-    provided 'com.segment.analytics.android:analytics:4.0.0'
+    provided 'com.segment.analytics.android:analytics:4.+'
 
     testCompile 'junit:junit:4.12'
     testCompile 'org.mockito:mockito-core:1.10.19'
 
-    testCompile 'com.segment.analytics.android:analytics-tests:4.0.0'
+    testCompile 'com.segment.analytics.android:analytics-tests:4.+'
 
 }
