+++ app/src/main/java/com/segment/analytics/android/integrations/appsflyer/AppsflyerIntegration.java
@@ -61,6 +61,7 @@ public class AppsflyerIntegration extends Integration<AppsFlyerLib> {
             String devKey = settings.getString("appsFlyerDevKey");
             boolean trackAttributionData = settings.getBoolean("trackAttributionData", false);
             Application application = analytics.getApplication();
+            afLib.trackEvent(application.getApplicationContext(), null, null);
             afLib.startTracking(application, devKey);
             logger.verbose("AppsFlyer.getInstance().startTracking(%s, %s)", application, devKey.substring(0, 1) + "*****" + devKey.substring(devKey.length() - 2));
             if (trackAttributionData) {
@@ -85,15 +86,19 @@ public class AppsflyerIntegration extends Integration<AppsFlyerLib> {
         this.isDebug = (logger.logLevel != Analytics.LogLevel.NONE);
     }
 
+    // TODO: for some reason it never called
     @Override
     public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
         super.onActivityCreated(activity, savedInstanceState);
+
         updateEndUserAttributes();
         appsflyer.startTracking(activity.getApplication(), appsFlyerDevKey);
         logger.verbose("AppsFlyer.getInstance().startTracking(%s, %s)",
                 activity.getApplication(),
                 appsFlyerDevKey.substring(0, 1) + "*****" + appsFlyerDevKey.substring(appsFlyerDevKey.length() - 2));
     }
+
+
 
     @Override
     public AppsFlyerLib getUnderlyingInstance() {
