+++ segmenttestapp/src/main/AndroidManifest.xml
@@ -10,7 +10,9 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
+        android:theme="@style/AppTheme"
+        android:name=".SampleApplication"
+        >
 
         <activity android:name=".MainActivity">
             <intent-filter>
